Optimizes the way we process changes (from changes feed)
Instead of listening to the change feed, we fetch 100 changes then push them to the queue. Once the queue is drained, we fetch the next 100 changes and so on. When there are no new changes in the feed we switch to listening to changes. As soon as a new change is found, we switch back to fetching (100 at a time).